redux사용 및 파일분리
