+++ src/states/index.ts
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit'
 import sidebar from './sidebar'
+import totalAd from './totalAd'
 
 export const store = configureStore({
   reducer: {
     sidebar,
+    totalAd,
   },
   devTools: process.env.NODE_ENV !== 'production',
   middleware: (getDefaultMiddleware) => getDefaultMiddleware({ serializableCheck: false }),
